+++ pism_terra/glacier/run.py
@@ -36,9 +36,9 @@ from jinja2 import Environment, FileSystemLoader, StrictUndefined
 from pyfiglet import Figlet
 
 from pism_terra.aws import local_to_s3
-from pism_terra.climate import create_offset_file
 from pism_terra.config import JobConfig, RunConfig, load_config, load_uq
 from pism_terra.download import file_localizer
+from pism_terra.glacier.climate import create_offset_file
 from pism_terra.glacier.execute import find_first_and_execute
 from pism_terra.glacier.stage import stage_glacier
 from pism_terra.sampling import create_samples
@@ -184,6 +184,8 @@ def run_glacier(
     log_path.mkdir(parents=True, exist_ok=True)
     output_path = glacier_path / Path("output")
     output_path.mkdir(parents=True, exist_ok=True)
+    scalar_path = output_path / Path("scalar")
+    scalar_path.mkdir(parents=True, exist_ok=True)
     spatial_path = output_path / Path("spatial")
     spatial_path.mkdir(parents=True, exist_ok=True)
     state_path = output_path / Path("state")
@@ -239,11 +241,13 @@ def run_glacier(
     # Apply to runtime dict (these should be dotted PISM flags)
     run.update(overrides)
 
+    scalar_file = scalar_path / Path(f"scalar_g{resolution}_{rgi_id}_{name_options}_{start}_{end}.nc")
     spatial_file = spatial_path / Path(f"spatial_g{resolution}_{rgi_id}_{name_options}_{start}_{end}.nc")
     state_file = state_path / Path(f"state_g{resolution}_{rgi_id}_{name_options}_{start}_{end}.nc")
     run.update(
         {
             "output.file": state_file.resolve(),
+            "output.scalar.file": scalar_file.resolve(),
             "output.spatial.file": spatial_file.resolve(),
         }
     )
@@ -278,6 +282,7 @@ def run_glacier(
         "rgi": {"rgi_id": rgi_id, "outline": str(outline_file.resolve())},
         "output": {
             "spatial": str(spatial_file.resolve()),
+            "scalar.file": scalar_file.resolve(),
             "state": str(state_file.resolve()),
         },
         "config": run,
@@ -595,7 +600,8 @@ def run_ensemble():
         "surface.force_to_thickness.file": df["boot_file"].iloc[0],
         "atmosphere.delta_T.file": df["scalar_offset_file"].iloc[0],
         "atmosphere.elevation_change.file": df["climate_file"].iloc[0],
-        "atmosphere.fract_P.file": df["scalar_offset_file"].iloc[0],
+        "atmosphere.frac_P.file": df["scalar_offset_file"].iloc[0],
+        "atmosphere.precip_scaling.file": df["scalar_offset_file"].iloc[0],
         "atmosphere.given.file": df["climate_file"].iloc[0],
     }
     outline_file = df["outline"].iloc[0]
@@ -634,6 +640,7 @@ def run_ensemble():
         frac_P = row["atmosphere.frac_P"] if "atmosphere.frac_P" in row else 0
         create_offset_file(scalar_offset_file, delta_T=delta_T, frac_P=frac_P)
         row["atmosphere.delta_T.file"] = scalar_offset_file
+        row["atmosphere.frac_P.file"] = scalar_offset_file
         row["atmosphere.precip_scaling.file"] = scalar_offset_file
         run_glacier(
             rgi_id,